Workaround issue #50.
The macros advice-add and add-function were added in Emacs 24.4.
There's no reason to use old style defadvice with add-function.  This
change disables special wrappers that handle indenting of 'static if',
'version' and 'debug' constructs in d-mode for versions of Emacs prior
to version 24.4 because the macros that install the wrappers were not
available until the version 24.4.