Fix indentation on Emacs 26
In Emacs 26, cc-mode adds a new concept (noise macros) along with
several related constants and variables. Unless these are properly
processed during cc-mode initialization, attempting to indent code will
result in the error: (wrong-type-argument stringp nil)

This commit adds an invocation to the appropriate initialization
function when it is present.