Move c-forward-decl-or-cast-1 into d-mode.el
c-forward-decl-or-cast-1 is very complicated and tries to do a lot of
things at once. Due to this, we are reaching the limit of what we can
do with only non-invasive patching.

Maintaining a copy (as d-forward-decl-or-cast-1) will allow us to
implement support for D syntax which does not align with the logic in
cc-mode's implementation, remove code for complicated rules which are
not relevant to D syntax, and consolidate the ad-hoc patches
previously applied using advice.